[#154072850] specs fixed
